+++ packages/react-magma-dom/src/components/DatePicker/DatePicker.test.js
@@ -491,7 +491,11 @@ describe('Date Picker', () => {
       });
 
       expect(
-        getByText(subWeeks(defaultDate, 1).getDate().toString())
+        getByText(
+          subWeeks(defaultDate, 1)
+            .getDate()
+            .toString()
+        )
       ).not.toHaveStyleRule('border-color', 'transparent');
     });
 
@@ -512,7 +516,11 @@ describe('Date Picker', () => {
       });
 
       expect(
-        getByText(subDays(defaultDate, 1).getDate().toString())
+        getByText(
+          subDays(defaultDate, 1)
+            .getDate()
+            .toString()
+        )
       ).not.toHaveStyleRule('border-color', 'transparent');
     });
 
@@ -533,7 +541,11 @@ describe('Date Picker', () => {
       });
 
       expect(
-        getByText(startOfWeek(defaultDate, 1).getDate().toString())
+        getByText(
+          startOfWeek(defaultDate, 1)
+            .getDate()
+            .toString()
+        )
       ).not.toHaveStyleRule('border-color', 'transparent');
     });
 
@@ -554,7 +566,11 @@ describe('Date Picker', () => {
       });
 
       expect(
-        getByText(subMonths(defaultDate, 1).getDate().toString())
+        getByText(
+          subMonths(defaultDate, 1)
+            .getDate()
+            .toString()
+        )
       ).not.toHaveStyleRule('border-color', 'transparent');
     });
 
@@ -575,7 +591,11 @@ describe('Date Picker', () => {
       });
 
       expect(
-        getByText(addMonths(defaultDate, 1).getDate().toString())
+        getByText(
+          addMonths(defaultDate, 1)
+            .getDate()
+            .toString()
+        )
       ).not.toHaveStyleRule('border-color', 'transparent');
     });
 
@@ -596,7 +616,11 @@ describe('Date Picker', () => {
       });
 
       expect(
-        getByText(addWeeks(defaultDate, 1).getDate().toString())
+        getByText(
+          addWeeks(defaultDate, 1)
+            .getDate()
+            .toString()
+        )
       ).not.toHaveStyleRule('border-color', 'transparent');
     });
 
@@ -617,7 +641,11 @@ describe('Date Picker', () => {
       });
 
       expect(
-        getByText(addDays(defaultDate, 1).getDate().toString())
+        getByText(
+          addDays(defaultDate, 1)
+            .getDate()
+            .toString()
+        )
       ).not.toHaveStyleRule('border-color', 'transparent');
     });
 
@@ -638,7 +666,11 @@ describe('Date Picker', () => {
       });
 
       expect(
-        getByText(endOfWeek(defaultDate).getDate().toString())
+        getByText(
+          endOfWeek(defaultDate)
+            .getDate()
+            .toString()
+        )
       ).not.toHaveStyleRule('border-color', 'transparent');
     });