+++ packages/react-magma-dom/src/components/Table/TableHeaderCell.tsx
@@ -78,10 +78,10 @@ const StyledTableHeaderCell = styled.th<{
     `}
 
     ${props =>
-    props.width &&
-    css`
-      width: ${props.width};
-    `}
+      props.width &&
+      css`
+        width: ${props.width};
+      `}
 `;
 
 const SortButton = styled.button<{